#2884: wip
