+++ factcast-core/src/main/java/org/factcast/core/snap/SnapshotCache.java
@@ -32,6 +32,4 @@ public interface SnapshotCache {
   void setSnapshot(@NonNull Snapshot snapshot);
 
   void clearSnapshot(@NonNull SnapshotId id);
-
-  void compact(int retentionTimeInDays);
 }
