+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/server/grpc/FactCastGrpcServerAutoConfiguration.java
@@ -15,7 +15,6 @@
  */
 package org.factcast.spring.boot.autoconfigure.server.grpc;
 
-import lombok.Generated;
 import net.devh.boot.grpc.server.autoconfigure.*;
 import org.factcast.server.grpc.FactCastGrpcServerConfiguration;
 import org.factcast.server.grpc.FactCastSecurityConfiguration;
@@ -26,7 +25,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Import;
 
-@Generated
 @AutoConfiguration
 @Import({FactCastGrpcServerConfiguration.class, FactCastSecurityConfiguration.class})
 @ConditionalOnClass(FactStoreGrpcService.class)
