+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/server/grpc/LZ4ServerAutoConfiguration.java
@@ -15,7 +15,6 @@
  */
 package org.factcast.spring.boot.autoconfigure.server.grpc;
 
-import lombok.Generated;
 import net.jpountz.lz4.LZ4Compressor;
 import org.factcast.grpc.lz4.Lz4GrpcServerCodec;
 import org.factcast.server.grpc.FactStoreGrpcService;
@@ -28,7 +27,6 @@ import org.springframework.context.annotation.Bean;
  * @deprecated in 0.8
  */
 @SuppressWarnings("DeprecatedIsStillUsed")
-@Generated
 @AutoConfiguration
 @ConditionalOnClass({FactStoreGrpcService.class, LZ4Compressor.class, Lz4GrpcServerCodec.class})
 @Deprecated
