+++ factcast-spring-boot-autoconfigure/src/main/java/org/factcast/spring/boot/autoconfigure/server/ui/FactCastServerUIAutoConfiguration.java
@@ -15,13 +15,11 @@
  */
 package org.factcast.spring.boot.autoconfigure.server.ui;
 
-import lombok.Generated;
 import org.factcast.server.ui.config.UIConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Import;
 
-@Generated
 @AutoConfiguration
 @Import({UIConfiguration.class})
 @ConditionalOnClass(UIConfiguration.class)
