+++ factcast-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,11 +1,11 @@
 org.factcast.spring.boot.autoconfigure.core.FactCastAutoConfiguration
-org.factcast.spring.boot.autoconfigure.core.FactCastSnapshotCacheAutoConfiguration
-org.factcast.spring.boot.autoconfigure.local.snap.InMemorySnapshotCacheAutoConfiguration
-org.factcast.spring.boot.autoconfigure.local.snap.InMemoryAndDiskSnapshotCacheAutoConfiguration
+org.factcast.spring.boot.autoconfigure.snap.InMemorySnapshotCacheAutoConfiguration
+org.factcast.spring.boot.autoconfigure.snap.InMemoryAndDiskSnapshotCacheAutoConfiguration
 org.factcast.spring.boot.autoconfigure.redis.RedissonSnapshotCacheAutoConfiguration
 org.factcast.spring.boot.autoconfigure.redis.RedissonAutoConfiguration
 org.factcast.spring.boot.autoconfigure.factus.FactusAutoConfiguration
 org.factcast.spring.boot.autoconfigure.factus.BinaryJacksonSnapshotSerializerAutoConfiguration
+org.factcast.spring.boot.autoconfigure.snap.FactCastSnapshotCacheAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.GrpcFactStoreAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.LZ4ClientAutoConfiguration
 org.factcast.spring.boot.autoconfigure.client.grpc.SnappyClientAutoConfiguration
